# generated/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/tasks": {
      "post": {
        "operationId": "TasksController_create",
        "summary": "Создать новую задачу",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Задача успешно создана",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Неверные данные"
          },
          "401": {
            "description": "Не авторизован"
          }
        },
        "tags": [
          "tasks"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "get": {
        "operationId": "TasksController_findAll",
        "summary": "Get tasks for the current user",
        "description": "Returns tasks authored by the current user. Supports optional filters: status (filter), project scope (projectId), and keyword search on description (search).",
        "parameters": [
          {
            "name": "filter",
            "required": false,
            "in": "query",
            "description": "Filter by completion status",
            "schema": {
              "enum": [
                "current",
                "past"
              ],
              "type": "string"
            }
          },
          {
            "name": "projectId",
            "required": false,
            "in": "query",
            "description": "Scope results to a specific project",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Case-insensitive keyword search on task description",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Task list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskResponseDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "tasks"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/tasks/meeting/{meetingId}": {
      "get": {
        "operationId": "TasksController_findByMeeting",
        "summary": "Получить все задачи из конкретной встречи",
        "parameters": [
          {
            "name": "meetingId",
            "required": true,
            "in": "path",
            "description": "ID встречи",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Список задач встречи",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskResponseDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Неверный ID встречи"
          },
          "401": {
            "description": "Не авторизован"
          }
        },
        "tags": [
          "tasks"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/tasks/{id}": {
      "get": {
        "operationId": "TasksController_findOne",
        "summary": "Получить задачу по ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID задачи",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Детали задачи",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Не авторизован"
          },
          "403": {
            "description": "Можно просматривать только свои задачи"
          },
          "404": {
            "description": "Задача не найдена"
          }
        },
        "tags": [
          "tasks"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "patch": {
        "operationId": "TasksController_update",
        "summary": "Обновить задачу (только автор)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID задачи",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Задача обновлена",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Не авторизован"
          },
          "403": {
            "description": "Только автор может обновить задачу"
          }
        },
        "tags": [
          "tasks"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "delete": {
        "operationId": "TasksController_remove",
        "summary": "Удалить задачу (только автор)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID задачи",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Задача удалена"
          },
          "401": {
            "description": "Не авторизован"
          },
          "403": {
            "description": "Только автор может удалить задачу"
          }
        },
        "tags": [
          "tasks"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/tasks/{id}/approve": {
      "patch": {
        "operationId": "TasksController_approveTask",
        "summary": "Approve or unapprove task (creator only)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID задачи",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveTaskDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Task approval status updated"
          },
          "401": {
            "description": "Не авторизован"
          },
          "403": {
            "description": "Only meeting creator can approve tasks"
          },
          "404": {
            "description": "Task or meeting not found"
          }
        },
        "tags": [
          "tasks"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "Meetings Quality API",
    "description": "API для платформы отслеживания качества встреч",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateTaskDto": {
        "type": "object",
        "properties": {
          "projectId": {
            "type": "string",
            "description": "Project this task belongs to",
            "example": "507f1f77bcf86cd799439050"
          },
          "description": {
            "type": "string",
            "description": "Описание задачи",
            "example": "Реализовать аутентификацию пользователей"
          },
          "commonQuestion": {
            "type": "string",
            "description": "Описание общего вопроса",
            "example": "Как я понимаю общий вопрос "
          },
          "meetingId": {
            "type": "string",
            "description": "ID встречи, из которой создана задача",
            "example": "507f1f77bcf86cd799439020"
          },
          "deadline": {
            "type": "string",
            "description": "Дедлайн задачи (ISO формат)",
            "example": "2026-02-01T00:00:00.000Z"
          },
          "estimateHours": {
            "type": "number",
            "description": "Оценить время на задачу",
            "example": "2"
          },
          "contributionImportance": {
            "type": "number",
            "description": "Важность вклада (0-100)",
            "example": 90,
            "minimum": 0,
            "maximum": 100
          }
        },
        "required": [
          "description",
          "commonQuestion",
          "meetingId",
          "deadline",
          "estimateHours",
          "contributionImportance"
        ]
      },
      "TaskAuthorRefDto": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "507f1f77bcf86cd799439011"
          },
          "fullName": {
            "type": "string",
            "example": "Jane Doe",
            "nullable": true
          },
          "email": {
            "type": "string",
            "example": "jane@example.com",
            "nullable": true
          }
        },
        "required": [
          "_id",
          "fullName",
          "email"
        ]
      },
      "TaskMeetingRefDto": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "507f1f77bcf86cd799439020"
          },
          "title": {
            "type": "string",
            "example": "Q3 Planning Session"
          },
          "question": {
            "type": "string",
            "example": "What are the priorities this quarter?"
          }
        },
        "required": [
          "_id",
          "title",
          "question"
        ]
      },
      "TaskResponseDto": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "507f1f77bcf86cd799439030"
          },
          "description": {
            "type": "string",
            "example": "Implement authentication module"
          },
          "commonQuestion": {
            "type": "string",
            "example": "How should we approach the tech stack?"
          },
          "authorId": {
            "$ref": "#/components/schemas/TaskAuthorRefDto"
          },
          "meetingId": {
            "$ref": "#/components/schemas/TaskMeetingRefDto"
          },
          "deadline": {
            "format": "date-time",
            "type": "string",
            "example": "2026-02-01T00:00:00.000Z"
          },
          "estimateHours": {
            "type": "number",
            "example": 8,
            "description": "Estimated hours to complete the task"
          },
          "contributionImportance": {
            "type": "number",
            "example": 80,
            "minimum": 0,
            "maximum": 100
          },
          "approved": {
            "type": "boolean",
            "example": false
          },
          "isCompleted": {
            "type": "boolean",
            "example": false
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-01-19T10:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-01-19T10:00:00.000Z"
          }
        },
        "required": [
          "_id",
          "description",
          "commonQuestion",
          "authorId",
          "meetingId",
          "deadline",
          "estimateHours",
          "contributionImportance",
          "approved",
          "isCompleted",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateTaskDto": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string",
            "description": "Описание задачи",
            "example": "Обновленное описание задачи"
          },
          "deadline": {
            "type": "string",
            "description": "Дедлайн задачи (ISO формат)",
            "example": "2026-02-15T00:00:00.000Z"
          },
          "estimateHours": {
            "type": "number",
            "description": "Оценить время на задачу",
            "example": "2"
          },
          "contributionImportance": {
            "type": "number",
            "description": "Важность вклада (0-100)",
            "example": 95,
            "minimum": 0,
            "maximum": 100
          },
          "isCompleted": {
            "type": "boolean",
            "description": "Статус завершения задачи",
            "example": true
          }
        },
        "required": [
          "estimateHours"
        ]
      },
      "ApproveTaskDto": {
        "type": "object",
        "properties": {
          "approved": {
            "type": "boolean",
            "description": "Approve or unapprove the task",
            "example": true
          }
        },
        "required": [
          "approved"
        ]
      }
    }
  }
}